Move shared components to app/shared
