+++ app/features/home/HomePage.tsx
@@ -2,7 +2,7 @@
 import React, { useState, useMemo } from 'react';
 import { useTheme } from '@/app/providers/ThemeContext';
 import Link from 'next/link';
-import { Search, Sun, Moon } from '@/app/components/shared/SvgIcons';
+import { Search, Sun, Moon } from '@/app/shared/SvgIcons';
 import surahsData from '@/data/surahs.json';
 import juzData from '@/data/juz.json';
 import VerseOfDay from './VerseOfDay';
