+++ app/features/player/components/AudioSettingsModal.tsx
@@ -1,6 +1,6 @@
 'use client';
 import { useState } from 'react';
-import { FaTimes } from '@/app/components/shared/SvgIcons';
+import { FaTimes } from '@/app/shared/SvgIcons';
 import { useAudio, RepeatOptions } from '@/app/features/player/context/AudioContext';
 import { RECITERS } from '@/lib/audio/reciters';
 import { useTranslation } from 'react-i18next';
